+++ packages/core/test/Core.spec.js
@@ -269,8 +269,8 @@ describe("Core", () => {
     it("should add Route Handler", () => {
       core.addRoutesHandler("foo");
       // TODO, do not use private properties in testing
-      expect(core._variantHandlers._registeredVariantHandlers.length).toEqual(4);
-      expect(core._variantHandlers._registeredVariantHandlers[3]).toEqual("foo");
+      expect(core._variantHandlers._registeredVariantHandlers.length).toEqual(6);
+      expect(core._variantHandlers._registeredVariantHandlers[5]).toEqual("foo");
     });
   });