+++ test/core-e2e-legacy/src/load-mocks-and-routes.spec.js
@@ -56,7 +56,7 @@ describe("loadMocks and loadRoutes methods", () => {
         {
           id: "get-books",
           url: "/api/books",
-          method: "GET",
+          method: "get",
           variants: [
             {
               id: "success",
@@ -79,7 +79,7 @@ describe("loadMocks and loadRoutes methods", () => {
         {
           id: "get-authors",
           url: "/api/authors",
-          method: "GET",
+          method: "get",
           variants: [
             {
               id: "success",